requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.97.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.27.8
pydantic==1.10.9
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
yarl==1.9.2
